+++ src/main/java/net/coralmod/mod/render/TextureRenderer.java
@@ -5,8 +5,8 @@ import net.minecraft.world.item.Item;
 
 public interface TextureRenderer {
 
-    void draw(Identifier texture, float x, float y, int width, int height);
+    void draw(Identifier texture, int x, int y, int width, int height);
 
-    void item(Item item, float x, float y, float scale);
+    void item(Item item, int x, int y, float scale);
 
 }
